Removed mongodb stuff. maybe later. playlists are now per user. Static playlists are now supported.

+++ mysql/playlist.go
@@ -7,11 +7,34 @@ import (
 	"go2music/model"
 )
 
+const (
+	createPlaylistTableStatement = `
+	CREATE TABLE IF NOT EXISTS playlist (
+		id varchar(32),
+		name VARCHAR(255) NOT NULL,
+		query VARCHAR(255) NOT NULL,
+		user_id varchar(32) NOT NULL,
+		PRIMARY KEY (id),
+		FOREIGN KEY (user_id) REFERENCES user(id)
+		);
+	`
+
+	createPlaylistSongTableStatement = `
+	CREATE TABLE IF NOT EXISTS playlist_song (
+		playlist_id varchar(32) NOT NULL,
+		song_id varchar(32) NOT NULL,
+		PRIMARY KEY (playlist_id,song_id),
+		FOREIGN KEY (playlist_id) REFERENCES playlist(id),
+		FOREIGN KEY (song_id) REFERENCES song(id)
+		);
+	`
+)
+
 func (db *DB) initializePlaylist() {
 	_, err := db.Query("SELECT 1 FROM playlist LIMIT 1")
 	if err != nil {
 		log.Info("Table playlist does not exists. Creating now.")
-		stmt, err := db.Prepare("CREATE TABLE IF NOT EXISTS playlist (id varchar(32), name varchar(255) NOT NULL, query varchar(255) NOT NULL, PRIMARY KEY (id));")
+		stmt, err := db.Prepare(createPlaylistTableStatement)
 		if err != nil {
 			log.Error("Error creating playlist table")
 			panic(fmt.Sprintf("%v", err))
@@ -35,12 +58,28 @@ func (db *DB) initializePlaylist() {
 		} else {
 			log.Info("Index on name generated....")
 		}
+		_, err = db.Query("SELECT 1 FROM playlist_song LIMIT 1")
+		if err != nil {
+			log.Info("Table playlist_song does not exists. Creating now.")
+			stmt, err := db.Prepare(createPlaylistSongTableStatement)
+			if err != nil {
+				log.Error("Error creating playlist_song table")
+				panic(fmt.Sprintf("%v", err))
+			}
+			_, err = stmt.Exec()
+			if err != nil {
+				log.Error("Error creating playlist_song table")
+				panic(fmt.Sprintf("%v", err))
+			} else {
+				log.Info("Playlist_song Table successfully created....")
+			}
+		}
 	}
 }
 
 func (db *DB) CreatePlaylist(playlist model.Playlist) (*model.Playlist, error) {
 	playlist.Id = xid.New().String()
-	_, err := db.Exec("INSERT IGNORE INTO playlist (id,name,query) VALUES(?,?,?)", playlist.Id, playlist.Name, playlist.Query)
+	_, err := db.Exec("INSERT IGNORE INTO playlist (id,name,query,user_id) VALUES(?,?,?,?)", playlist.Id, playlist.Name, playlist.Query, playlist.User.Id)
 	if err != nil {
 		log.Fatal(err)
 	}
@@ -48,12 +87,12 @@ func (db *DB) CreatePlaylist(playlist model.Playlist) (*model.Playlist, error) {
 }
 
 func (db *DB) CreateIfNotExistsPlaylist(playlist model.Playlist) (*model.Playlist, error) {
-	existingPlaylist, findErr := db.FindPlaylistByName(playlist.Name)
+	existingPlaylist, findErr := db.FindPlaylistByName(playlist.Name, playlist.User.Id)
 	if findErr == nil {
 		return existingPlaylist, findErr
 	}
 	playlist.Id = xid.New().String()
-	_, err := db.Exec("INSERT INTO playlist (id,name,query) VALUES(?,?,?)", playlist.Id, playlist.Name, playlist.Query)
+	_, err := db.Exec("INSERT INTO playlist (id,name,query,user_id) VALUES(?,?,?,?)", playlist.Id, playlist.Name, playlist.Query, playlist.User.Id)
 	if err != nil {
 		log.Fatal(err)
 	}
@@ -65,31 +104,34 @@ func (db *DB) UpdatePlaylist(playlist model.Playlist) (*model.Playlist, error) {
 	return &playlist, err
 }
 
-func (db *DB) DeletePlaylist(id string) error {
-	_, err := db.Exec("DELETE FROM playlist WHERE id=?", id)
+func (db *DB) DeletePlaylist(id string, user_id string) error {
+	_, err := db.Exec("DELETE FROM playlist_song WHERE playlist_id=?", id)
+	if err == nil {
+		_, err = db.Exec("DELETE FROM playlist WHERE id=? AND user_id=?", id, user_id)
+	}
 	return err
 }
 
-func (db *DB) FindPlaylistById(id string) (*model.Playlist, error) {
+func (db *DB) FindPlaylistById(id string, user_id string) (*model.Playlist, error) {
 	playlist := new(model.Playlist)
-	err := db.QueryRow("SELECT id,name,query FROM playlist WHERE id=?", id).Scan(&playlist.Id, &playlist.Name, &playlist.Query)
+	err := db.QueryRow("SELECT id,name,query FROM playlist WHERE id=? AND user_id=?", id, user_id).Scan(&playlist.Id, &playlist.Name, &playlist.Query)
 	if err != nil {
 		log.Fatal(err)
 	}
 	return playlist, err
 }
 
-func (db *DB) FindPlaylistByName(name string) (*model.Playlist, error) {
+func (db *DB) FindPlaylistByName(name string, user_id string) (*model.Playlist, error) {
 	playlist := new(model.Playlist)
-	err := db.QueryRow("SELECT id,name,query FROM playlist WHERE name=?", name).Scan(&playlist.Id, &playlist.Name, &playlist.Query)
+	err := db.QueryRow("SELECT id,name,query FROM playlist WHERE name=? AND user_id=?", name, user_id).Scan(&playlist.Id, &playlist.Name, &playlist.Query)
 	if err != nil {
 		return playlist, err
 	}
 	return playlist, err
 }
 
-func (db *DB) FindAllPlaylists() ([]*model.Playlist, error) {
-	rows, err := db.Query("SELECT id, name, query FROM playlist")
+func (db *DB) FindAllPlaylists(user_id string) ([]*model.Playlist, error) {
+	rows, err := db.Query("SELECT id, name, query FROM playlist WHERE user_id=?", user_id)
 	if err != nil {
 		log.Fatal(err)
 	}
